simplify test/wait
