+++ src/MPI.jl
@@ -46,6 +46,7 @@ include("environment.jl")
 include("datatypes.jl")
 include("buffers.jl")
 include("operators.jl")
+include("nonblocking.jl")
 include("pointtopoint.jl")
 include("collective.jl")
 include("topology.jl")
