+++ test/test_wait.jl
@@ -26,14 +26,14 @@ send_check = zeros(Int, nsends)
 recv_check = zeros(Int, nsends)
 
 for i = 1:nsends
-    idx, stat = MPI.Waitany!(send_reqs)
+    idx = MPI.Waitany(send_reqs)
     @test send_reqs[idx] == MPI.REQUEST_NULL
     send_check[idx] += 1
 end
 @test send_check == ones(Int, nsends)
 
 for i = 1:nsends
-    idx, stat = MPI.Waitany!(recv_reqs)
+    idx = MPI.Waitany(recv_reqs)
     @test recv_reqs[idx] == MPI.REQUEST_NULL
     recv_check[idx] += 1
 end
